+++ .gitignore
@@ -1,4 +1,1 @@
-.idea
-.vscode
 node_modules
-.coverage
