+++ src/__tests__/hooks/easy.useCalendarView.spec.ts
@@ -2,7 +2,6 @@ import { act, renderHook } from '@testing-library/react';
 import { vi } from 'vitest';
 
 import { useCalendarView } from '../../hooks/useCalendarView.ts';
-import { assertDate } from '../utils.ts';
 
 describe('초기 상태', () => {
   it('view는 "month"이어야 한다', () => {
