+++ src/__tests__/unit/easy.eventOverlap.spec.ts
@@ -1,4 +1,3 @@
-import { Event } from '../../types';
 import {
   convertEventToDateRange,
   findOverlappingEvents,
