+++ src/__tests__/unit/easy.eventUtils.spec.ts
@@ -1,4 +1,3 @@
-import { Event } from '../../types';
 import { getFilteredEvents } from '../../utils/eventUtils';
 
 describe('getFilteredEvents', () => {
