+++ src/components/Notifications.tsx
@@ -1,0 +1,29 @@
+import { Alert, AlertIcon, AlertTitle, Box, CloseButton, VStack } from '@chakra-ui/react';
+
+type Notification = {
+  message: string;
+};
+
+type Props = {
+  notifications: Notification[];
+  // eslint-disable-next-line no-unused-vars
+  onClose: (index: number) => void;
+};
+
+export const Notifications = ({ notifications, onClose }: Props) => {
+  if (notifications.length === 0) return null;
+
+  return (
+    <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
+      {notifications.map((notification, index) => (
+        <Alert key={index} status="info" variant="solid" width="auto">
+          <AlertIcon />
+          <Box flex="1">
+            <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
+          </Box>
+          <CloseButton onClick={() => onClose(index)} />
+        </Alert>
+      ))}
+    </VStack>
+  );
+};
